+++ Static/home1.html
@@ -1,56 +1,3 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8" />
-  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-  <title>RaidX</title>
-  <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet" />
-  <style>
-    body {
-      background-color: #f3f6f5;
-    }
-    .hero {
-      text-align: center;
-      margin-top: 50px;
-    }
-    .hero h1 span.orange {
-      color: #f97316;
-    }
-    .hero h1 span.blue {
-      color: #2563eb;
-    }
-    .btn-orange {
-      background-color: #f97316;
-      color: white;
-    }
-    .btn-orange:hover {
-      background-color: #ea580c;
-      color: white;
-    }
-    .info-box {
-      background-color: #f8f9fa;
-      border-radius: 12px;
-      padding: 30px;
-      margin-top: 40px;
-      box-shadow: 0 2px 8px rgba(0,0,0,0.05);
-    }
-    .info-box h5 {
-      font-weight: 600;
-    }
-    .profile {
-      display: flex;
-      align-items: center;
-      gap: 10px;
-    }
-    .profile img {
-      width: 40px;
-      height: 40px;
-      border-radius: 50%;
-    }
-  </style>
-</head>
-<body>
-
   <!-- Top Right Profile Section -->
 <!DOCTYPE html>
 <html lang="en">
