+++ internal/handlers/playerprofile.go
@@ -12,6 +12,17 @@ import (
 	"go.mongodb.org/mongo-driver/mongo"
 )
 
+var PlayerProfile struct {
+	FullName      string    `bson:"fullName"`
+	Email         string    `bson:"email"`
+	UserId        string    `bson:"userId"`
+	Position      string    `bson:"position"`
+	CreatedAt     time.Time `bson:"createdAt"`
+	TotalPoints   int       `bson:"totalPoints"`
+	RaidPoints    int       `bson:"raidPoints"`
+	DefencePoints int       `bson:"defencePoints"`
+}
+
 func PlayerProfileHandler(c *fiber.Ctx) error {
 	// Get the player ID from the URL
 	playerID := c.Params("id")
@@ -27,20 +38,10 @@ func PlayerProfileHandler(c *fiber.Ctx) error {
 	collection := db.MongoClient.Database("raidx").Collection("players")
 
 	// Query to find the player by ID
-	var player struct {
-		FullName      string    `bson:"fullName"`
-		Email         string    `bson:"email"`
-		UserId        string    `bson:"userId"`
-		Position      string    `bson:"position"`
-		CreatedAt     time.Time `bson:"createdAt"`
-		TotalPoints   int       `bson:"totalPoints"`
-		RaidPoints    int       `bson:"raidPoints"`
-		DefencePoints int       `bson:"defencePoints"`
-	}
 
 	// Find the player by ID
 	filter := bson.M{"_id": objID}
-	err = collection.FindOne(context.TODO(), filter).Decode(&player)
+	err = collection.FindOne(context.TODO(), filter).Decode(&PlayerProfile)
 	if err != nil {
 		if err == mongo.ErrNoDocuments {
 			logrus.Info("Info:", "PlayerProfileHandler:", " Player not found: %v", err)
@@ -53,13 +54,13 @@ func PlayerProfileHandler(c *fiber.Ctx) error {
 	// Render the player profile HTML with the data
 	return c.Render("playerprofile", fiber.Map{
 		"ID":            playerID,
-		"FullName":      player.FullName,
-		"Email":         player.Email,
-		"UserId":        player.UserId,
-		"Position":      player.Position,
-		"CreatedAt":     player.CreatedAt.Format("2006-01-02"), // Format for readability
-		"TotalPoints":   player.TotalPoints,
-		"RaidPoints":    player.RaidPoints,
-		"DefencePoints": player.DefencePoints,
+		"FullName":      PlayerProfile.FullName,
+		"Email":         PlayerProfile.Email,
+		"UserId":        PlayerProfile.UserId,
+		"Position":      PlayerProfile.Position,
+		"CreatedAt":     PlayerProfile.CreatedAt.Format("2006-01-02"), // Format for readability
+		"TotalPoints":   PlayerProfile.TotalPoints,
+		"RaidPoints":    PlayerProfile.RaidPoints,
+		"DefencePoints": PlayerProfile.DefencePoints,
 	})
 }
